06-research-agent/requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
